requirements.txt
discord.py
black
